# docs/api/admin.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "admin-service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/agents/{agent_id}/runs": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agent Runs",
        "description": "Return the most recent ``agent_runs`` rows for one agent.",
        "operationId": "list_agent_runs_admin_agents__agent_id__runs_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/agents/{agent_id}/run": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Trigger Agent Run",
        "description": "Forward to ``agent-runtime`` and audit log the trigger.",
        "operationId": "trigger_agent_run_admin_agents__agent_id__run_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/agents/{agent_id}/constitution": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Constitution",
        "description": "Return the constitution markdown for an agent.",
        "operationId": "get_agent_constitution_admin_agents__agent_id__constitution_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConstitutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentConstitutionResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "constitution_path": {
            "type": "string",
            "title": "Constitution Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "constitution_path",
          "content"
        ],
        "title": "AgentConstitutionResponse",
        "description": "``GET /admin/agents/{id}/constitution`` payload."
      },
...
      "AgentMessageDTO": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "instrument_symbol": {
            "type": "string",
            "title": "Instrument Symbol"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "instrument_symbol",
          "decision",
          "rationale"
        ],
        "title": "AgentMessageDTO",
        "description": "Peer-agent rationale forwarded to agent-runtime."
      },
      "AgentRunRow": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "triggered_by": {
            "type": "string",
            "title": "Triggered By"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Input Tokens"
          },
          "total_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Output Tokens"
          },
          "total_cost_usd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost Usd"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "triggered_by",
          "started_at",
          "status"
        ],
        "title": "AgentRunRow",
        "description": "One row from ``agent_runs`` for the runs listing."
      },
      "AgentRunsResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/AgentRunRow"
            },
            "type": "array",
            "title": "Runs"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "runs",
          "limit"
        ],
        "title": "AgentRunsResponse",
        "description": "``GET /admin/agents/{id}/runs`` payload."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunAgentRequest": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "kind": {
            "type": "string",
            "pattern": "^(run|rebuttal)$",
            "title": "Kind",
            "default": "run"
          },
          "agent_messages": {
            "items": {
              "$ref": "#/components/schemas/AgentMessageDTO"
            },
            "type": "array",
            "title": "Agent Messages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot_id"
        ],
        "title": "RunAgentRequest",
        "description": "``POST /admin/agents/{id}/run`` body \u2014 forwarded to agent-runtime."
      },
      "RunAgentResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "run"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "snapshot_id"
        ],
        "title": "RunAgentResponse",
        "description": "``POST /admin/agents/{id}/run`` proxied response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}